fixes according to mentor's comments
