+++ project/src/components/movie-page/action/get-reviews.js
@@ -1,0 +1,5 @@
+import {fetchReviewList} from '../../../store/api-actions';
+
+export const getReviews = () => (dispatch, id) => {
+  dispatch(fetchReviewList(id));
+};
